feat: Upload dictionaries to data folder for notebook
This commit includes the addition of multiple dictionaries that are essential for the operation of the notebook. The files are uploaded into the 'data' folder to maintain project structure and ensure easy accessibility. This also improves data management within the project.

- Added 'dictonary.txt' from https://inventwithpython.com/dictionary.txt
- Added 'words.txt' from https://greenteapress.com/thinkpython/code/words.txt
- Added 'wordlist.txt' from http://www-personal.umich.edu/~jlawler/wordlist
- Added '12dicts.zip'
- Added 'scowl.zip'

Please ensure the respective dictionaries are loaded in the notebook as per usage.

Note: Please check and validate the dictionary files before use to ensure data integrity.